+++ openfe/tests/utils/test_duecredit.py
@@ -20,6 +20,10 @@ class TestDuecredit:
         ['openfe.protocols.openmm_rfe.equil_rfe_methods',
          ['10.5281/zenodo.1297683', '10.5281/zenodo.596622',
           '10.1371/journal.pcbi.1005659']],
+        ['openfe.protocols.openmm_utils.multistate_analysis',
+         ["10.5281/zenodo.596622", "10.1063/1.2978177",
+          "10.1021/ct0502864", "10.1021/acs.jctc.5b00784",
+          "10.5281/zenodo.596220"]],
     ])
     def test_duecredit_protocol_collection(self, module, dois):
         importlib.import_module(module)